+++ cleo_plugins/Audio/C3DAudioStream.cpp
@@ -1,6 +1,7 @@
 #include "C3DAudioStream.h"
 #include "CSoundSystem.h"
 #include "CLEO_Utils.h"
+#include "CCamera.h"
 
 using namespace CLEO;
 
@@ -15,7 +16,7 @@ C3DAudioStream::C3DAudioStream(const char* filepath) : CAudioStream()
         return;
     }
 
-    unsigned flags = BASS_SAMPLE_3D | BASS_SAMPLE_MONO | BASS_SAMPLE_SOFTWARE;
+    unsigned flags = BASS_SAMPLE_3D | BASS_SAMPLE_MONO | BASS_SAMPLE_SOFTWARE | BASS_STREAM_PRESCAN;
     if (CSoundSystem::useFloatAudio) flags |= BASS_SAMPLE_FLOAT;
 
     if (!(streamInternal = BASS_StreamCreateFile(FALSE, filepath, 0, 0, flags)) &&
@@ -26,7 +27,7 @@ C3DAudioStream::C3DAudioStream(const char* filepath) : CAudioStream()
     }
 
     BASS_ChannelGetAttribute(streamInternal, BASS_ATTRIB_FREQ, &rate);
-    BASS_ChannelSet3DAttributes(streamInternal, BASS_3DMODE_NORMAL, 3.0f, 1E+12f, -1, -1, -1.0f);
+    BASS_ChannelSet3DAttributes(streamInternal, BASS_3DMODE_NORMAL, -1.0f, -1.0f, -1, -1, -1.0f);
     BASS_ChannelSetAttribute(streamInternal, BASS_ATTRIB_VOL, 0.0f); // muted until processed
     ok = true;
 }
@@ -40,7 +41,7 @@ void C3DAudioStream::Set3dPosition(const CVector& pos)
 
 void C3DAudioStream::Set3dSourceSize(float radius)
 {
-    BASS_ChannelSet3DAttributes(streamInternal, BASS_3DMODE_NORMAL, radius, 1E+12f, -1, -1, -1.0f);
+    this->radius = std::max<float>(radius, 0.01f);
 }
 
 void C3DAudioStream::SetHost(CEntity* host, const CVector& offset)
@@ -61,8 +62,100 @@ void C3DAudioStream::SetHost(CEntity* host, const CVector& offset)
 
 void C3DAudioStream::Process()
 {
-    CAudioStream::Process();
     UpdatePosition();
+    CAudioStream::Process();
+
+    // position and velocity
+    CVector relPos = position - CSoundSystem::position;
+    float distance = relPos.NormaliseAndMag();
+    float inFactor = (float)CalculateDistanceDecay(radius * 5.0f, distance * 5.0f); // use decay curve for blending inside-outside source effects
+
+    // stereo panning
+    float sign = dot(CSoundSystem::direction, relPos) > 0.0f ? 1.0f : -1.0f;
+    CVector centerPos = CSoundSystem::position + CSoundSystem::direction * distance * sign;
+    CVector percPos = lerp(position, centerPos, inFactor);
+
+    CVector percVel = lerp(velocity, CSoundSystem::velocity, inFactor);
+
+    BASS_ChannelSet3DPosition(streamInternal, &toBass(percPos), nullptr, &toBass(percVel));
+}
+
+float C3DAudioStream::CalculateVolume()
+{
+    if (!placed)
+    {
+        return 0.0f; // mute until processed
+    }
+
+    CVector relPos = position - CSoundSystem::position;
+    float distance = relPos.NormaliseAndMag();
+    float inFactor = (float)CalculateDistanceDecay(radius * 5.0f, distance * 5.0f); // use decay curve for blending inside-outside source effects
+
+    double vol = Volume_3D_Adjust;
+
+    switch (type)
+    {
+        case SoundEffect: vol *= CSoundSystem::masterVolumeSfx; break;
+        case Music: vol *= CSoundSystem::masterVolumeMusic; break;
+        case UserInterface: vol *= CSoundSystem::masterVolumeSfx; break;
+        default: vol *= 1.0f; break;
+    }
+
+    // distance decay
+    vol *= CalculateDistanceDecay(radius, distance);
+
+    // "look" direction decay
+    vol *= lerp(CalculateDirectionDecay(CSoundSystem::direction, relPos), 1.0f, inFactor);
+
+    // screen black fade
+    if (type != UserInterface && !TheCamera.m_bIgnoreFadingStuffForMusic)
+    {
+        vol *= 1.0f - TheCamera.m_fFadeAlpha / 255.0f;
+    }
+
+    // music volume lowering in cutscenes, when characters talk, mission sounds are played etc.
+    if (type == Music) 
+    {
+        if (TheCamera.m_bWideScreenOn) vol *= 0.25f;
+    }
+
+    // stream's volume
+    vol *= volume.value();
+
+    return (float)vol;
+}
+
+float C3DAudioStream::CalculateSpeed()
+{
+    float masterSpeed;
+    switch (type)
+    {
+        case SoundEffect: masterSpeed = CSoundSystem::masterSpeed; break;
+        case Music: masterSpeed = CSoundSystem::masterSpeed; break;
+        case UserInterface: masterSpeed = 1.0f; break;
+        default: masterSpeed = 1.0f;
+    }
+
+    return masterSpeed * speed.value();
+}
+
+double C3DAudioStream::CalculateDistanceDecay(float radius, float distance)
+{
+    distance = std::max<float>(distance - radius, 0.0f);
+
+    // exact match to ingame sounds
+    /*float factor = 1.0f / powf(1.0f + distance, 2); // inverse square
+    factor -= 0.00006f;
+    factor = std::clamp(factor, 0.0f, 1.0f);*/
+
+    return exp(-0.017 * pow(distance, 1.3)); // more natural feeling
+}
+
+float C3DAudioStream::CalculateDirectionDecay(const CVector& listenerDir, const CVector& relativePos)
+{
+    float factor = dot(listenerDir, relativePos);
+    factor = 0.6f + 0.4f * factor; // 0.2 to 1.0
+    return factor;
 }
 
 void C3DAudioStream::UpdatePosition()
@@ -92,6 +185,7 @@ void C3DAudioStream::UpdatePosition()
         if (!hostValid)
         {
             hostType = ENTITY_TYPE_NOTHING;
+            placed = false;
             Stop();
             return;
         }
@@ -105,13 +199,12 @@ void C3DAudioStream::UpdatePosition()
 
     if (prevPos.Magnitude() > 0.0f) // not equal to 0,0,0
     {
-        CVector velocity = position - prevPos;
+        velocity = position - prevPos;
         velocity /= CSoundSystem::timeStep;
-        BASS_ChannelSet3DPosition(streamInternal, &toBass(position), nullptr, &toBass(velocity));
+        placed = true;
     }
     else
     {
-        BASS_ChannelSetAttribute(streamInternal, BASS_ATTRIB_VOL, 0.0f); // muted until next update
+        placed = false;
     }
 }
-